MultiComponentFlash

Documentation for MultiComponentFlash.

MultiComponentFlash.AbstractFlash

MultiComponentFlash.AbstractGeneralizedCubic

MultiComponentFlash.AbstractNewtonFlash

MultiComponentFlash.FlashedMixture2Phase

MultiComponentFlash.FlashedPhase

MultiComponentFlash.GenericCubicEOS

MultiComponentFlash.GenericCubicEOS

MultiComponentFlash.MolecularProperty

MultiComponentFlash.MolecularProperty

MultiComponentFlash.MultiComponentMixture

MultiComponentFlash.NewtonFlash

MultiComponentFlash.PengRobinson

MultiComponentFlash.RedlichKwong

MultiComponentFlash.SSIFlash

MultiComponentFlash.SSINewtonFlash

MultiComponentFlash.SinglePhaseLiquid

MultiComponentFlash.SinglePhaseVapor

MultiComponentFlash.SoaveRedlichKwong

MultiComponentFlash.TwoPhaseLiquidVapor

MultiComponentFlash.UnknownPhaseState

MultiComponentFlash.ZudkevitchJoffe

MultiComponentFlash.acentric_factor

MultiComponentFlash.component_fugacity

MultiComponentFlash.component_fugacity_coefficient

MultiComponentFlash.critical_pressure

MultiComponentFlash.critical_temperature

MultiComponentFlash.critical_volume

MultiComponentFlash.cubic_benchmark

MultiComponentFlash.cubic_polynomial

MultiComponentFlash.eos_polynomial

MultiComponentFlash.eos_polynomial

MultiComponentFlash.flash_2ph

MultiComponentFlash.flash_2ph!

MultiComponentFlash.flash_storage

MultiComponentFlash.force_coefficients

MultiComponentFlash.force_coefficients!

MultiComponentFlash.force_scalars

MultiComponentFlash.initial_guess_K

MultiComponentFlash.initial_guess_K!

MultiComponentFlash.inverse_flash_update!

MultiComponentFlash.lbc_viscosities

MultiComponentFlash.lbc_viscosity

MultiComponentFlash.liquid_mole_fraction

MultiComponentFlash.mass_densities

MultiComponentFlash.mass_density

MultiComponentFlash.mixture_fugacities

MultiComponentFlash.mixture_fugacities!

MultiComponentFlash.molar_volume

MultiComponentFlash.molar_weight

MultiComponentFlash.number_of_components

MultiComponentFlash.number_of_components

MultiComponentFlash.phase_is_present

MultiComponentFlash.phase_saturations

MultiComponentFlash.set_partials

MultiComponentFlash.set_partials_phase_mole_fractions!

MultiComponentFlash.set_partials_vapor_fraction

MultiComponentFlash.single_phase_label

MultiComponentFlash.solve_rachford_rice

MultiComponentFlash.solve_roots

MultiComponentFlash.update_attractive_linear!

MultiComponentFlash.update_attractive_quadratic!

MultiComponentFlash.vapor_mole_fraction

MultiComponentFlash.vapor_mole_fraction

MultiComponentFlash.wilson_estimate

MultiComponentFlash.wilson_estimate

MultiComponentFlash.wilson_estimate!

MultiComponentFlash.AbstractFlash — Type

Abstract type for all flash types

MultiComponentFlash.AbstractGeneralizedCubic — Type

Abstract base type for generalized cubics.

Any subtype may override the any of following internal functions to alter the EOS:

* static_coefficients
* weight_ai
* weight_bi

In addition, the GenericCubicEOS is parametric on the specific cubic type for further dispatch modifications.

MultiComponentFlash.AbstractNewtonFlash — Type

Abstract type for all flash types that use Newton in some form

MultiComponentFlash.FlashedMixture2Phase — Type

Type that holds liquid and vapor phase states together with their state

MultiComponentFlash.FlashedPhase — Type

Type that holds values for a flashed phase (mole fractions + compressibility factor)

MultiComponentFlash.GenericCubicEOS — Type

GenericCubicEOS is an implementation of generalized cubic equations of state.

Many popular cubic equations can be written in a single form with a few changes in definitions for the terms (they are, after all, all cubic in form). References:

Cubic Equations of State-Which? by J.J. Martin

Simulation of Gas Condensate Reservoir Performance by K.H. Coats

MultiComponentFlash.GenericCubicEOS — Type

GenericCubicEOS(mixture, [type = PengRobinson()])

Instantiate a generic cubic equation-of-state for a MultiComponentMixture and a specified EOS.

Currently supported choices for type:

1. `PengRobinson` (default)
2. `ZudkevitchJoffe`
3. `RedlichKwong`
4. `SoaveRedlichKwong`

MultiComponentFlash.MolecularProperty — Type

MolecularProperty(molar_mass, p_crit, T_crit, V_crit, acentric_factor = 0.0)

Type that defines the static properties of a molecular species.

MultiComponentFlash.MolecularProperty — Method

MolecularProperty("Name")

Convenience constructor that looks up molecular properties from a table in tabulated_properties.

The properties are taken from the wonderful MIT-licensed CoolProp. Please note that the equations of state included in this module may not be approprioate for all the available fluids, especially for mixtures!

See list of species at CoolProp website.

MultiComponentFlash.MultiComponentMixture — Type

MultiComponentMixture(properties::NTuple{N, MolecularProperty}; A_ij = nothing, names = ["C1", "C2", ...], name = "UnnamedMixture")

Create a multicomponent mixture with an optional binary interaction coefficient matrix A_ij.

MultiComponentFlash.NewtonFlash — Type

Flash using Newton's method for zero solve.

Only conditionally convergent, but has better convergence rate than SSI.

See also: flash_2ph!, SSIFlash SSINewtonFlash

MultiComponentFlash.PengRobinson — Type

Specializes the GenericCubicEOS to the Peng-Robinson cubic equation of state.

MultiComponentFlash.RedlichKwong — Type

Specializes the GenericCubicEOS to the Redlich-Kwong cubic equation of state.

MultiComponentFlash.SSIFlash — Type

Flash method that uses successive subtition.

Unconditionally convergent, does not require derivatives, but is very slow around critical regions.

See also: flash_2ph!, NewtonFlash SSINewtonFlash

MultiComponentFlash.SSINewtonFlash — Type

SSINewtonFlash([swap_iter = 5])

Perform a number of SSI iterations, followed by Newton until convergence.

See also: flash_2ph!, SSIFlash NewtonFlash

MultiComponentFlash.SinglePhaseLiquid — Type

Single-phase liquid state for dispatch

MultiComponentFlash.SinglePhaseVapor — Type

Single-phase vapor state for dispatch

MultiComponentFlash.SoaveRedlichKwong — Type

Specializes the GenericCubicEOS to the Soave-Redlich-Kwong cubic equation of state.

MultiComponentFlash.TwoPhaseLiquidVapor — Type

Two-phase liquid-vapor state for dispatch

MultiComponentFlash.UnknownPhaseState — Type

Unknown phase state (not initialized)

MultiComponentFlash.ZudkevitchJoffe — Type

Specializes the GenericCubicEOS to the Zudkevitch-Joffe cubic equation of state.

The Zudkevitch-Joffe equations of state allows for per-component functions of temperature that modify the weight_ai and weight_bi functions. These additional fitting parameters allows for more flexibility when matching complex mixtures.

MultiComponentFlash.acentric_factor — Method

Get the acentric factorfor a species.

MultiComponentFlash.component_fugacity — Method

component_fugacity(eos, cond, i, Z, forces, scalars)

Get fugacity of component i in a phase with compressibility Z and EOS constants scalars.

MultiComponentFlash.component_fugacity_coefficient — Method

Fugacity

MultiComponentFlash.critical_pressure — Method

Get the critical pressure for a species.

MultiComponentFlash.critical_temperature — Method

Get the critical temperature for a species.

MultiComponentFlash.critical_volume — Method

Get the critical volume for a species.

MultiComponentFlash.cubic_benchmark — Method

cubic_benchmark(name)

Get a benchmark equation-of-state instance together with some data for testing.

MultiComponentFlash.cubic_polynomial — Method

(Precompute version)

MultiComponentFlash.eos_polynomial — Method

Generic version (only depend on scalars, ignores forces)

MultiComponentFlash.eos_polynomial — Method

(Expensive version)

MultiComponentFlash.flash_2ph! — Function

flash_2ph!(storage, eos, c, K; <keyword arguments>)

Non-allocating version of flash_2ph where storage is pre-allocated.

Useful if you are performing many flashes of the same system with varying conditions.

Arguments

storage: Should be output from flash_storage(eos, c, method = method). Preallocated storage.

Remaining arguments documented in flash_2ph.

Keyword arguments

update_forces = true: Update the p, T dependent forces in storage initially.

MultiComponentFlash.flash_2ph — Method

flash_2ph(eos, c, [K]; <keyword arguments>)

Perform two-phase flash with a given EOS under a set of specific conditions. Returns vapor fraction. Modifies K in-place.

Given a mixture with pressure, temperature and mole fractions, this routine performs a vapor-liquid equilibrium calculation after a stability test.

Two outcomes are possible:

A single-phase condition prevails (returned vapor fraction is NaN) and a single phase (liquid or vapor) is stable.

A two-phase condition is possible. The routine produces K-values and vapor fraction so that the following holds:

Isofugacity constraint for all components ($f_{li} = f_{vi}$)

Molar balance for all components ($(1-V) x_i - V y_i - z_i$)

Unity condition ($\sum_i (x_i - y_i) = 0$)

Arguments

eos: the equation-of-state to be used for the flash

c: conditions to flash the mixture at on the form (p = 10e5, T = 303.15, z = [0.5, 0.3, 0.2])

K: optionally a buffer of length number_of_components(eos) used to hold K-values. Modified in-place.

Keyword arguments

method = SSIFlash(): Flash method to use. Can be SSIFlash(), NewtonFlash() or SSINewtonFlash().

tolerance = 1e-8: Tolerance for the convergence criterion. Relative to $\|1-R_i\|_\infty$ where $R_i = f_{il}/f_{iv}$

maxiter = 10000: Maximum nubmer of iterations for both stability tests and the main flash.

verbose = false: Emit extra information during solve.

extra_out = false: Return (V, K, conv) where conv contains iterations and oncergence status instead of just V.

See also: flash_2ph!, single_phase_label

MultiComponentFlash.flash_storage — Function

flash_storage(eos, [c]; <keyword arguments>)

Pre-allocate storage for flash_2ph!.

Arguments

eos: the equation-of-state to be used for the flash

c: conditions for the mixture. Types in conditions must match later usage (e.g. for use with ForwardDiff).

Keyword arguments

method = SSIFlash(): Flash method to use. Can be SSIFlash(), NewtonFlash() or SSINewtonFlash().

static_size = false: Use SArrays and MArrays for fast flash, but slower compile times.

inc_jac: Allocate storage for Newton/Jacobian. Required for Newton (and defaults to true for that method) or for diff_externals.

diff_externals = false: Allocate storage for matrix inversion required to produce partial derivatives of flash using set_partials.

See also: flash_2ph! set_partials

MultiComponentFlash.force_coefficients! — Method

force_coefficients!(coeff, eos, cond)

In-place update of force coefficients.

See also force_coefficients

MultiComponentFlash.force_coefficients — Method

force_coefficients(eos, cond)

Get coefficients for forces for a specific EOS (component interactions). For most cubics, these are a set of attractive (linear and quadratic) forces and a set of linear repulsive forces.

Note that the current implementation of flash assumes that these are independent of the compositions themselves.

See also force_coefficients!

MultiComponentFlash.force_scalars — Method

force_scalars(eos, cond, forces)

Compute EOS specific scalars for the current conditions based on the forces.

MultiComponentFlash.initial_guess_K! — Method

initial_guess_K(eos, cond)

In-place version of initial_guess_K.

MultiComponentFlash.initial_guess_K — Method

initial_guess_K(eos, cond)

Produce a plausible initial guess for K values for eos under current cond.

MultiComponentFlash.inverse_flash_update! — Method

inverse_flash_update!(storage, eos, c, V)

Update internal matrix of partial derivatives for a converged flash result.

MultiComponentFlash.lbc_viscosities — Method

lbc_viscosities(eos, p, T, flashed_mixture)

Compute phase viscosities for a flashed two-phase mixture using the LBC correlation.

Always returns a named tuple of (μl, μv), even if the mixture is single-phase.

The value in the absent phase will be zero.

MultiComponentFlash.lbc_viscosity — Method

lbc_viscosity(eos, p, T, ph; <keyword arguments>)

Compute the viscosity of a mixture using the Lohrenz-Bray-Clark correlation.

MultiComponentFlash.liquid_mole_fraction — Method

Compute liquid mole fraction from overall mole fraction, K-value and V vapor fraction

MultiComponentFlash.mass_densities — Method

mass_densities(eos, p, T, flashed_mixture)

Compute mass densities for a flashed two-phase mixture.

Always returns a named tuple of (ρl, ρv), even if the mixture is single-phase.

The value in the absent phase will be zero.

MultiComponentFlash.mass_density — Method

Compute mass density of a flashed phase

MultiComponentFlash.mixture_fugacities! — Function

Compute fugacities for all components.

MultiComponentFlash.mixture_fugacities — Method

Allocating version of mixture_fugacities!

MultiComponentFlash.molar_volume — Method

Compute molar volume of a flashed phase

MultiComponentFlash.molar_weight — Method

Get the molar weight for a species.

MultiComponentFlash.number_of_components — Method

number_of_components(eos)

Return number of components for the underlying mixture of the EOS.

MultiComponentFlash.number_of_components — Method

number_of_components(mixture)

Return number of components in the MultiComponentMixture.

MultiComponentFlash.phase_is_present — Method

phase_is_present(label, phase_state)

Check if a phase (symbol :liquid/:vapor) is present with the provided phase state.

MultiComponentFlash.phase_saturations — Method

phase_saturations(flashed_mixture)

Compute phase saturations for a flashed two-phase mixture.

Always returns a named tuple of (Sl, Sv), even if the mixture is single-phase.

The value in the absent phase will be zero.

MultiComponentFlash.set_partials — Method

set_partials(v, storage, eos, c, index)

Modify a value v of ForwardDiff.Dual type to get the correct derivatives.

If the mixture is two-phase, the partial derivatives of phase molar fractions and vapor fraction with respect to the flash conditions (pressure, temperature and overall mole fractions) are not trivial. This function is the main gateway for setting these values.

Notes

Experimental interface, subject to change. You most likely want to use either set_partials_phase_mole_fractions! or set_partials_vapor_fraction

In order for this routine to work, the storage must be initialized using flash_storage with the following options enabled:

inc_jac = true

diff_externals = true

and inverse_flash_update! must be called after a successful flash. The partial derivatives with respect to p, T, z is then contained in storage.buf_inv with a negative sign. This function then performs the requisite chain rule operations for the input.

MultiComponentFlash.set_partials_phase_mole_fractions! — Function

setpartialsphasemolefractions!(xy, storage, eos, ∂c, [phase_symbol])

Set partial derivatives to phase mole fraction vector with type ForwardDiff.Dual with correct values, but missing partial derivatives.

MultiComponentFlash.set_partials_vapor_fraction — Method

set_partials_vapor_fraction(V, storage, eos, ∂c)

Set partial derivatives to a vapor mole fraction ForwardDiff.Dual instance with correct value, but missing partial derivatives.

MultiComponentFlash.single_phase_label — Method

single_phase_label(mixture, cond)

Li's method for single-phase labeling of a mixture. Estimate of pure vapor/liquid.

Returns a vapor fraction that is either 1.0 (=pure vapor) or 0.0 (=pure liquid).

MultiComponentFlash.solve_rachford_rice — Function

solve_rachford_rice(K, z, [V]; <keyword arguments>)

Compute vapor mole fraction V for given equilibrium constants K and mole fractions z.

Arguments

K - Equal length to z, containing the equilibrium constants for each component. z - Mole fractions. Should sum up to unity.

Keyword arguments

tol = 1e-12 - Tolerance for solve. maxiter - Maximum number of iterations ad - Use automatic differentiation (ForwardDiff) instead of analytical gradient.

Examples

julia> solve_rachford_rice([0.5, 1.5], [0.3, 0.7])
0.8000000000000002

MultiComponentFlash.solve_roots — Method

Specialization of solve_rotos for cubics

MultiComponentFlash.update_attractive_linear! — Method

Update linear part of attractive term

MultiComponentFlash.update_attractive_quadratic! — Method

Update quadratic part of attractive term

MultiComponentFlash.vapor_mole_fraction — Method

Compute vapor mole fraction from overall mole fraction, K-value and V vapor fraction

MultiComponentFlash.vapor_mole_fraction — Method

Compute vapor mole fraction from liquid mole fraction and K-value

MultiComponentFlash.wilson_estimate! — Method

wilson_estimate!(K, properties, p, T)

Update a vector K in-place with K-values from wilson_estimate.

MultiComponentFlash.wilson_estimate — Method

wilson_estimate(properties, p, T)

Create vector of K-values that holds the wilson_estimate for each species.

MultiComponentFlash.wilson_estimate — Method

wilson_estimate(p, T, ω, p_c, T_c)

Estimate K-values for a given acentric factor ω and pressure and temperature at current and critical conditions.

Reference Vapor-Liquid Equilibrium. XI. A New Expression for the Excess Free Energy of Mixing by GM Wilson