Equations of state

MultiComponentFlash.GenericCubicEOS

MultiComponentFlash.GenericCubicEOS

MultiComponentFlash.PengRobinson

MultiComponentFlash.RedlichKwong

MultiComponentFlash.SoaveRedlichKwong

MultiComponentFlash.ZudkevitchJoffe

MultiComponentFlash.component_fugacity

MultiComponentFlash.force_coefficients

MultiComponentFlash.force_coefficients!

MultiComponentFlash.force_scalars

MultiComponentFlash.mixture_fugacities

MultiComponentFlash.number_of_components

Types

MultiComponentFlash.GenericCubicEOS — Type

GenericCubicEOS(mixture, [type = PengRobinson()])

Instantiate a generic cubic equation-of-state for a MultiComponentMixture and a specified EOS.

Currently supported choices for type:

1. `PengRobinson` (default)
2. `ZudkevitchJoffe`
3. `RedlichKwong`
4. `SoaveRedlichKwong`

MultiComponentFlash.GenericCubicEOS — Type

GenericCubicEOS is an implementation of generalized cubic equations of state.

Many popular cubic equations can be written in a single form with a few changes in definitions for the terms (they are, after all, all cubic in form). References:

Cubic Equations of State-Which? by J.J. Martin

Simulation of Gas Condensate Reservoir Performance by K.H. Coats

MultiComponentFlash.PengRobinson — Type

Specializes the GenericCubicEOS to the Peng-Robinson cubic equation of state.

MultiComponentFlash.RedlichKwong — Type

Specializes the GenericCubicEOS to the Redlich-Kwong cubic equation of state.

MultiComponentFlash.SoaveRedlichKwong — Type

Specializes the GenericCubicEOS to the Soave-Redlich-Kwong cubic equation of state.

MultiComponentFlash.ZudkevitchJoffe — Type

Specializes the GenericCubicEOS to the Zudkevitch-Joffe cubic equation of state.

The Zudkevitch-Joffe equations of state allows for per-component functions of temperature that modify the weight_ai and weight_bi functions. These additional fitting parameters allows for more flexibility when matching complex mixtures.

Functions

MultiComponentFlash.component_fugacity — Method

component_fugacity(eos, cond, i, Z, forces, scalars)

Get fugacity of component i in a phase with compressibility Z and EOS constants scalars.

MultiComponentFlash.force_coefficients! — Method

force_coefficients!(coeff, eos, cond)

In-place update of force coefficients.

See also force_coefficients

MultiComponentFlash.force_coefficients — Method

force_coefficients(eos, cond)

Get coefficients for forces for a specific EOS (component interactions). For most cubics, these are a set of attractive (linear and quadratic) forces and a set of linear repulsive forces.

Note that the current implementation of flash assumes that these are independent of the compositions themselves.

See also force_coefficients!

MultiComponentFlash.force_scalars — Method

force_scalars(eos, cond, forces)

Compute EOS specific scalars for the current conditions based on the forces.

MultiComponentFlash.mixture_fugacities — Method

Allocating version of mixture_fugacities!

MultiComponentFlash.number_of_components — Method

number_of_components(eos)

Return number of components for the underlying mixture of the EOS.