Utilities

MultiComponentFlash.inverse_flash_update!

MultiComponentFlash.lbc_viscosities

MultiComponentFlash.lbc_viscosity

MultiComponentFlash.mass_densities

MultiComponentFlash.mass_density

MultiComponentFlash.molar_volume

MultiComponentFlash.phase_is_present

MultiComponentFlash.phase_saturations

MultiComponentFlash.set_partials

MultiComponentFlash.set_partials_phase_mole_fractions!

MultiComponentFlash.set_partials_vapor_fraction

MultiComponentFlash.single_phase_label

Partial derivatives

Experimental features

Functions for obtaining partial derivatives of the flashed results. Please note that this is an experimental feature. Examples of usage are found in the unit tests.

MultiComponentFlash.inverse_flash_update! — Method

inverse_flash_update!(storage, eos, c, V)

Update internal matrix of partial derivatives for a converged flash result.

MultiComponentFlash.set_partials — Method

set_partials(v, storage, eos, c, index)

Modify a value v of ForwardDiff.Dual type to get the correct derivatives.

If the mixture is two-phase, the partial derivatives of phase molar fractions and vapor fraction with respect to the flash conditions (pressure, temperature and overall mole fractions) are not trivial. This function is the main gateway for setting these values.

Notes

Experimental interface, subject to change. You most likely want to use either set_partials_phase_mole_fractions! or set_partials_vapor_fraction

In order for this routine to work, the storage must be initialized using flash_storage with the following options enabled:

inc_jac = true

diff_externals = true

and inverse_flash_update! must be called after a successful flash. The partial derivatives with respect to p, T, z is then contained in storage.buf_inv with a negative sign. This function then performs the requisite chain rule operations for the input.

MultiComponentFlash.set_partials_phase_mole_fractions! — Function

setpartialsphasemolefractions!(xy, storage, eos, ∂c, [phase_symbol])

Set partial derivatives to phase mole fraction vector with type ForwardDiff.Dual with correct values, but missing partial derivatives.

MultiComponentFlash.set_partials_vapor_fraction — Method

set_partials_vapor_fraction(V, storage, eos, ∂c)

Set partial derivatives to a vapor mole fraction ForwardDiff.Dual instance with correct value, but missing partial derivatives.

Coupling utilities

Experimental features

Utilities for coupling flash to simulation codes. Please note that this is an experimental feature.

MultiComponentFlash.lbc_viscosities — Method

lbc_viscosities(eos, p, T, flashed_mixture)

Compute phase viscosities for a flashed two-phase mixture using the LBC correlation.

Always returns a named tuple of (μl, μv), even if the mixture is single-phase.

The value in the absent phase will be a tiny value (eps of the numeric type) to make division for mobilities etc. safe.

MultiComponentFlash.mass_densities — Method

mass_densities(eos, p, T, flashed_mixture)

Compute mass densities for a flashed two-phase mixture.

Always returns a named tuple of (ρl, ρv), even if the mixture is single-phase.

The value in the absent phase will be zero.

MultiComponentFlash.mass_density — Method

Compute mass density of a flashed phase

MultiComponentFlash.molar_volume — Method

Compute molar volume of a flashed phase

MultiComponentFlash.phase_is_present — Method

phase_is_present(label, phase_state)

Check if a phase (symbol :liquid/:vapor) is present with the provided phase state.

MultiComponentFlash.phase_saturations — Method

phase_saturations(eos, p, T, flashed_mixture)

Compute phase saturations for a flashed two-phase mixture.

Always returns a named tuple of (Sl, Sv), even if the mixture is single-phase.

The value in the absent phase will be zero.

Various

MultiComponentFlash.lbc_viscosity — Method

lbc_viscosity(eos, p, T, ph; <keyword arguments>)

Compute the viscosity of a mixture using the Lohrenz-Bray-Clark correlation.

MultiComponentFlash.single_phase_label — Method

single_phase_label(mixture, cond)

Li's method for single-phase labeling of a mixture. Estimate of pure vapor/liquid.

Returns a vapor fraction that is either 1.0 (=pure vapor) or 0.0 (=pure liquid).