Mixtures

MultiComponentFlash.MolecularProperty

MultiComponentFlash.MolecularProperty

MultiComponentFlash.MultiComponentMixture

MultiComponentFlash.acentric_factor

MultiComponentFlash.critical_pressure

MultiComponentFlash.critical_temperature

MultiComponentFlash.critical_volume

MultiComponentFlash.molar_weight

MultiComponentFlash.number_of_components

MultiComponentFlash.MolecularProperty — Type

MolecularProperty(molar_mass, p_crit, T_crit, V_crit, acentric_factor = 0.0)

Type that defines the static properties of a molecular species.

MultiComponentFlash.MolecularProperty — Method

MolecularProperty("Name")

Convenience constructor that looks up molecular properties from a table in tabulated_properties.

The properties are taken from the wonderful MIT-licensed CoolProp. Please note that the equations of state included in this module may not be approprioate for all the available fluids, especially for mixtures!

See list of species at CoolProp website.

MultiComponentFlash.MultiComponentMixture — Type

MultiComponentMixture(properties::NTuple{N, MolecularProperty}; A_ij = nothing, names = ["C1", "C2", ...], name = "UnnamedMixture")

Create a multicomponent mixture with an optional binary interaction coefficient matrix A_ij.

MultiComponentFlash.acentric_factor — Method

Get the acentric factorfor a species.

MultiComponentFlash.critical_pressure — Method

Get the critical pressure for a species.

MultiComponentFlash.critical_temperature — Method

Get the critical temperature for a species.

MultiComponentFlash.critical_volume — Method

Get the critical volume for a species.

MultiComponentFlash.molar_weight — Method

Get the molar weight for a species.

MultiComponentFlash.number_of_components — Method

number_of_components(mixture)

Return number of components in the MultiComponentMixture.